add simple user access rights

+++ todo/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.decorators import login_required
 from django.db.models import F
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseNotFound
 
 from todo.forms import ProjectForm, TaskForm, ProjectExecutorForm, SubTaskForm
 from todo.models import Project, Task, User, SubTask
@@ -25,19 +25,36 @@ def project_list(request):
 @login_required()
 def project_details(request, project_id):
     project = get_object_or_404(Project, pk=project_id)
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     tasks = Task.objects.filter(project_id=project_id)
+    progress = {}
+    for task in tasks:
+        progress[task.pk] = {
+            'total': SubTask.objects.filter(task_id=task.pk).count(),
+            'done': SubTask.objects.filter(task_id=task.pk, is_done=True).count(),
+        }
+        if progress[task.pk]['total'] == 0 or progress[task.pk]['done'] == 0:
+            progress[task.pk]['percent'] = 20
+        else:
+            progress[task.pk]['percent'] = int(progress[task.pk]['done'] / progress[task.pk]['total'] * 100)
+    print(progress)
     executors = list(project.executors.all())
     context = {
         'project': project,
         'tasks': tasks,
         'executors': executors,
+        'progress': progress,
     }
     return render(request, 'todo/project_details.html', context)
 
 
 @login_required()
 def task_create(request, **kwargs):
-    executors = Project.objects.get(pk=kwargs['project_id']).executors.all()
+    project = Project.objects.get(pk=kwargs['project_id'])
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
+    executors = project.executors.all()
     form_cont = tuple((user.pk, user.get_full_name()) for user in executors)
     form = TaskForm(request.POST or None, form_cont)
     if request.method == 'POST':
@@ -78,6 +95,8 @@ def project_create(request):
 def add_project_executor(request, project_id):
     form = ProjectExecutorForm(request.POST or None)
     project = Project.objects.get(pk=project_id)
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     context = {
         'form': form,
         'project_name': project.name,
@@ -94,6 +113,9 @@ def add_project_executor(request, project_id):
 
 
 def task_details(request, project_id, task_id):
+    project = Project.objects.get(pk=project_id)
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     task = get_object_or_404(Task, pk=task_id)
     subtasks = SubTask.objects.filter(task_id=task_id)
     context = {
@@ -104,6 +126,9 @@ def task_details(request, project_id, task_id):
 
 
 def subtask_create(request, **kwargs):
+    project = Project.objects.get(pk=kwargs['project_id'])
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     form = SubTaskForm(request.POST or None)
     if form.is_valid():
         subtask = form.save(commit=False)
@@ -120,12 +145,16 @@ def subtask_create(request, **kwargs):
 
 
 def subtask_done(request, project_id, task_id, subtask_id):
-    print(request, project_id, task_id, subtask_id)
+    project = Project.objects.get(pk=project_id)
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     SubTask.objects.filter(pk=subtask_id).update(is_done=(F('is_done') + 1) % 2)
     return redirect('todo:task_details', project_id=project_id, task_id=task_id)
 
 
 def subtask_delete(request, project_id, task_id, subtask_id):
-    print(request, project_id, task_id, subtask_id)
+    project = Project.objects.get(pk=project_id)
+    if request.user != project.author or request.user not in project.executors.all():
+        return HttpResponseNotFound('<h1>Page not found</h1>')
     SubTask.objects.filter(pk=subtask_id).delete()
     return redirect('todo:task_details', project_id=project_id, task_id=task_id)
